+++ src/tests/test_cases/noqa.py
@@ -1,4 +1,5 @@
-"""Test case for "# noqa" comments."""
+# noqa: D400,D415
+"""Test case for "# noqa" comments"""
 from .expected import Expectation