+++ docs/release_notes.rst
@@ -15,11 +15,22 @@ New Features
 
 * Extend support for detecting missing arguments in Google style
   docstrings to method calls (#384).
+* Extend support for detecting missing argument description in Numpy style
+  docstrings (#407).
+* Added support for Python 3.8 (#423).
+* Allow skipping errors on module level docstring via #noqa (#427).
 
 Bug Fixes
 
 * Remove D413 from the pep257 convention (#404).
 * Replace `semicolon` with `colon` in D416 messages. (#409)
+* D301 (Use r""" if any backslashes in a docstring) does not trigger on
+  backslashes for line continuation or unicode literals ``\u...`` and
+  ``\N...`` anymore. These are considered intended elements of the docstring
+  and thus should not be escaped by using a raw docstring (#365).
+* Fix decorator parsing (#411).
+* Google-style sections no longer cause false errors when used with
+  Numpy-style sections (#388, #424).
 * Fix D401 and D404 checks not working for docstrings containing only one word and ending with non-alpha character (#421)
 
 4.0.1 - August 14th, 2019
@@ -34,6 +45,7 @@ Bug Fixes
 * Fixed RST error in table which resulted in the online documentation missing
   the violation code table (#396).
 * Fixed IndentationError when parsing function arguments (#392).
+
 
 4.0.0 - July 6th, 2019
 ----------------------